[C] Rename 'AppSettings' forms and drawer to 'GlobalSettings'
