--- packages/admin/components/composed/settings/AppSettingsEditDrawer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppSettingsEditDrawer";
