--- packages/admin/components/composed/settings/AppSettingsEditForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppSettingsEditForm";
