+++ packages/admin/components/composed/settings/GlobalSettingsEditDrawer/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./GlobalSettingsEditDrawer";
