+++ packages/admin/components/composed/settings/GlobalSettingsEditForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./GlobalSettingsEditForm";
